Move getSubItems into TutorialRailItem.helpers.
